refactor(settings): move LDK debug UI
